Move prefill from ContactInfoPage to TypeOfCarePage
so the prefill happens when the user starts the New Appointment flow
